+++ src/block.h
@@ -8,5 +8,5 @@
 void ewsfs_block_read_size(FILE* file);
 void ewsfs_block_set_size(uint64_t block_size);
 uint64_t ewsfs_block_get_size();
-bool ewsfs_block_read(FILE* file, size_t block_index, uint8_t* buffer);
-bool ewsfs_block_write(FILE* file, size_t block_index, const uint8_t* buffer);
+bool ewsfs_block_read(FILE* file, uint64_t block_index, uint8_t* buffer);
+bool ewsfs_block_write(FILE* file, uint64_t block_index, const uint8_t* buffer);
